requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.4.2
starlette==0.27.0
aiohttp==3.8.6
python-multipart==0.0.6
httpx==0.25.0
pytest==7.4.3 